# public/openapi.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "PartFiles API",
		"description": "API documentation for PartFiles - a minimalist file management and sharing platform.",
		"version": "1.0.0"
	},
	"servers": [
		{
			"url": "/api",
			"description": "Main API"
		}
	],
...
	"paths": {
...
		"/content": {
			"get": {
				"tags": ["Admin", "Uploader"],
				"summary": "List content",
				"description": "Lists content. Admins may see all items (optional filters). Other roles must pass `uploadedBy` query params scoped to their username; the server enforces access.",
				"parameters": [
					{
						"name": "expired",
						"in": "query",
						"schema": {
							"type": "string",
							"enum": ["active", "expired"]
						}
					},
					{
						"name": "uploadedBy",
						"in": "query",
						"description": "Repeat for multiple tags (admin filter; required for non-admins).",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "List of content"
					},
					"400": {
						"description": "Invalid uploader filter"
					},
					"401": {
						"description": "Not authenticated"
					}
				}
			}
		},
...
	}
}